wallet_billing/requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
Django==5.2.5
django-environ==0.12.0
django-ninja==1.4.3
django-stubs==5.2.2
django-stubs-ext==5.2.2
factory_boy==3.3.3
Faker==37.6.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
sqlparse==0.5.3
types-PyYAML==6.0.12.20250822
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
